+++ projects/spine/src/test/java/org/openmuc/jeebus/spine/spi/FeaturePermissionTest.java
@@ -14,6 +14,7 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.function.ThrowingSupplier;
+import org.junit.jupiter.api.parallel.Execution;
 import org.openmuc.jeebus.spine.api.Device;
 import org.openmuc.jeebus.spine.api.Error;
 import org.openmuc.jeebus.spine.api.RequestResult;
@@ -23,8 +24,10 @@ import org.openmuc.jeebus.spine.xsd.v1.*;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
 
+import static org.junit.jupiter.api.parallel.ExecutionMode.SAME_THREAD;
 import static org.openmuc.jeebus.spine.TestUtilities.*;
 
+@Execution(SAME_THREAD)
 class FeaturePermissionTest {
 
     @BeforeEach